+++ src/orquestra/sdk/_base/_driver/_models.py
@@ -201,9 +201,10 @@ class WorkflowRunResponse(MinimalWorkflowRunResponse):
 class Resources(pydantic.BaseModel):
     """
     Implements:
-        https://github.com/zapatacomputing/workflow-driver/blob/6270a214fff40f53d7b25ec967f2e7875eb296e3/openapi/src/schemas/Resources.yaml
+        https://github.com/zapatacomputing/workflow-driver/blob/580c8d8/openapi/src/schemas/Resources.yaml
     """
 
+    nodes: Optional[int]
     cpu: Optional[str] = pydantic.Field(
         regex=r"^([+-]?[0-9.]+)([eEinumkKMGTP]*[-+]?[0-9]*)$"
     )