+++ src/orquestra/sdk/_base/_workflow.py
@@ -69,6 +69,7 @@ class WorkflowDef(Generic[_R]):
         name: str,
         workflow_fn: Callable[..., _R],
         fn_ref: FunctionRef,
+        resources: _dsl.Resources,
         data_aggregation: Optional[DataAggregation] = None,
         workflow_args: Optional[Tuple[Any, ...]] = None,
         workflow_kwargs: Optional[Dict[str, Any]] = None,
@@ -78,6 +79,7 @@ class WorkflowDef(Generic[_R]):
         self._name = name
         self._fn = workflow_fn
         self._fn_ref = fn_ref
+        self._resources = resources
         self._data_aggregation = data_aggregation
         self._workflow_args = workflow_args or ()
         self._workflow_kwargs = workflow_kwargs or {}
@@ -256,6 +258,54 @@ class WorkflowDef(Generic[_R]):
         run.start()
         return run
 
+    def with_resources(
+        self,
+        *,
+        cpu: Optional[Union[str, _dsl.Sentinel]] = _dsl.Sentinel.NO_UPDATE,
+        memory: Optional[Union[str, _dsl.Sentinel]] = _dsl.Sentinel.NO_UPDATE,
+        disk: Optional[Union[str, _dsl.Sentinel]] = _dsl.Sentinel.NO_UPDATE,
+        gpu: Optional[Union[str, _dsl.Sentinel]] = _dsl.Sentinel.NO_UPDATE,
+        nodes: Optional[Union[int, _dsl.Sentinel]] = _dsl.Sentinel.NO_UPDATE,
+    ) -> "WorkflowDef":
+        """
+        Assigns optional metadata related to this workflow definition object.
+
+        Doesn't modify the existing workflow definition, returns a new one.
+
+        Example usage:
+            wf_run = my_workflow().with_resources(
+                cpu="10", memory="10Gi"
+            ).run("my_cluster")
+
+        Args:
+            cpu: amount of cpu requested for the workflow
+            memory: amount of memory requested for the workflow
+            disk: amount of disk requested for the workflow
+            gpu: amount of gpu requested for the workflow
+            nodes: the number of nodes requested for the workflow
+        """
+        # Only use the new properties if they have not been changed.
+        # None is a valid option, so we are using the Sentinel object pattern:
+        # https://python-patterns.guide/python/sentinel-object/
+
+        resources = self._resources
+        new_resources = _dsl.Resources(
+            cpu=resources.cpu if cpu is _dsl.Sentinel.NO_UPDATE else cpu,
+            gpu=resources.gpu if gpu is _dsl.Sentinel.NO_UPDATE else gpu,
+            memory=resources.memory if memory is _dsl.Sentinel.NO_UPDATE else memory,
+            disk=resources.disk if disk is _dsl.Sentinel.NO_UPDATE else disk,
+            nodes=resources.nodes if nodes is _dsl.Sentinel.NO_UPDATE else nodes,
+        )
+        return WorkflowDef(
+            name=self._name,
+            workflow_fn=self._fn,
+            fn_ref=self._fn_ref,
+            resources=new_resources,
+            data_aggregation=self._data_aggregation,
+            workflow_args=self._workflow_args,
+            workflow_kwargs=self._workflow_kwargs,
+        )
+
 
 class WorkflowTemplate(Generic[_P, _R]):
     """
@@ -268,6 +318,7 @@ class WorkflowTemplate(Generic[_P, _R]):
         workflow_fn: Callable[_P, _R],
         fn_ref: FunctionRef,
         is_parametrized: bool,
+        resources: _dsl.Resources,
         data_aggregation: Optional[Union[DataAggregation, bool]] = None,
         default_source_import: Optional[Import] = None,
         default_dependency_imports: Optional[Iterable[Import]] = None,
@@ -275,8 +326,9 @@ class WorkflowTemplate(Generic[_P, _R]):
         self._custom_name = custom_name
         self._fn = workflow_fn
         self._fn_ref = fn_ref
-        self._data_aggregation = data_aggregation
         self._is_parametrized = is_parametrized
+        self._resources = resources
+        self._data_aggregation = data_aggregation
         self._default_source_import = default_source_import
         self._default_dependency_imports = default_dependency_imports
 
@@ -351,14 +403,15 @@ class WorkflowTemplate(Generic[_P, _R]):
                     "Workflow arguments must be known at submission time. "
                 )
         return WorkflowDef(
-            name,
-            self._fn,
-            self._fn_ref,
-            data_aggregation,
-            args,
-            kwargs,
-            self._default_source_import,
-            self._default_dependency_imports,
+            name=name,
+            workflow_fn=self._fn,
+            fn_ref=self._fn_ref,
+            resources=self._resources,
+            data_aggregation=data_aggregation,
+            workflow_args=args,
+            workflow_kwargs=kwargs,
+            default_source_import=self._default_source_import,
+            default_dependency_imports=self._default_dependency_imports,
         )
 
     @property
@@ -531,6 +584,7 @@ def workflow(fn: Callable[_P, _R]) -> WorkflowTemplate[_P, _R]:
 @overload
 def workflow(
     *,
+    resources: Optional[_dsl.Resources] = None,
     data_aggregation: Optional[Union[DataAggregation, bool]] = None,
     custom_name: Optional[str] = None,
     default_source_import: Optional[Import] = None,
@@ -542,6 +596,7 @@ def workflow(
 def workflow(
     fn: Optional[Callable[_P, _R]] = None,
     *,
+    resources: Optional[_dsl.Resources] = None,
     data_aggregation: Optional[Union[DataAggregation, bool]] = None,
     custom_name: Optional[str] = None,
     default_source_import: Optional[Import] = None,
@@ -553,6 +608,12 @@ def workflow(
     """Decorator that produces a workflow definition.
 
     Args:
+        resources: !Unstable API! The resources that this workflow requires.
+            The exact behaviour depends on the runtime, based on a Compute Engine
+            workflow, you can set the cluster your workflow will use:
+            10 nodes with 20 CPUs and a GPU each would be:
+            resources=sdk.Resources(cpu="20", gpu="1", nodes=10)
+            If omitted, the cluster's default resources will be used.
         data_aggregation: Used to set up resources used during data step. If skipped,
             or assigned True default values will be used. If assigned False
             data aggregation step will not run.
@@ -604,6 +665,7 @@ def workflow(
         fn_ref = get_fn_ref(fn)
         template = WorkflowTemplate(
             custom_name=name,
+            resources=resources or _dsl.Resources(),
             workflow_fn=fn,
             fn_ref=fn_ref,
             is_parametrized=len(signature.parameters) > 0,